Move to redactions.go
