+++ server.go
@@ -31,9 +31,7 @@ type subscription struct {
 	fn func(*pqs.Event) bool
 }
 
-// RedactionFields describes how redaction fields are specified.
-// Top level map key is the schema, inner map key is the table and slice is the fields to redact.
-type RedactionFields map[string]map[string][]string
+
 
 // Server implements PQStreamServer and manages both client connections and database event monitoring.
 type Server struct {
@@ -62,12 +60,7 @@ func WithTableRegexp(re *regexp.Regexp) ServerOption {
 	}
 }
 
-// WithFieldRedactions controls which fields are redacted from the feed.
-func WithFieldRedactions(r RedactionFields) ServerOption {
-	return func(s *Server) {
-		s.redactions = r
-	}
-}
+
 
 // NewServer prepares a new pqstream server.
 func NewServer(connectionString string, opts ...ServerOption) (*Server, error) {
@@ -189,20 +182,7 @@ func (s *Server) fallbackLookup(e *pqs.Event) error {
 	return nil
 }
 
-// redactFields search through redactionMap if there's any redacted fields
-// specified that match the fields of the current event.
-func (s *Server) redactFields(e *pqs.Event) {
-	if tables, ok := s.redactions[e.GetSchema()]; ok {
-		if fields, ok := tables[e.GetTable()]; ok {
-			for _, rf := range fields {
-				if _, ok := e.Payload.Fields[rf]; ok {
-					//remove field from payload
-					delete(e.Payload.Fields, rf)
-				}
-			}
-		}
-	}
-}
+
 
 // HandleEvents processes events from the database and copies them to relevent clients.
 func (s *Server) HandleEvents(ctx context.Context) error {
